Add files via upload
The HA: Forensics challenge is a Capture the Flag (CTF) focused on digital forensics, where participants are tasked with finding four hidden flags. This challenge is available on VulnHub and provides an excellent opportunity to develop and showcase digital forensics expertise. To solve the challenge, I used Kali Linux and an OVA file (Ubuntu VM) as the primary tools. A detailed walkthrough and flag solutions can be found in the uploaded file. By completing this challenge, individuals can gain hands-on experience with digital forensics tools and techniques, preparing them for real-world investigations. If you'd like to contribute or suggest improvements, feel free to open an issue or submit a pull request. This challenge is a great resource for anyone looking to improve their digital forensics skills and learn new techniques for analyzing and solving CTF challenges. The HA: Forensics challenge is a valuable tool for developing expertise in digital forensics and is highly recommended for those interested in the field.